refactor: remove the logic that fetch from client
