+++ packages/error-overlay/src/constants.ts
