+++ packages/error-overlay/src/middleware/helper/createOriginalStackFrame.ts
