+++ packages/error-overlay/src/middleware/helper/findOriginalSourcePositionAndContent.ts
