+++ packages/error-overlay/src/middleware/helper/getSourceById.ts
@@ -3,8 +3,8 @@ import { RawSourceMap } from 'source-map';
 import dataUriToBuffer, { MimeBuffer } from 'data-uri-to-buffer';
 import type webpack from '@shuvi/toolpack/lib/webpack';
 
-import { getSourceMapUrl } from '../../shared/helper/getSourceMapUrl';
-import { getModuleById } from '../../shared/helper/getModuleById';
+import { getSourceMapUrl } from './getSourceMapUrl';
+import { getModuleById } from './getModuleById';
 
 export type Source = RawSourceMap | null;
