+++ packages/error-overlay/src/view/container/RuntimeError.tsx
@@ -18,7 +18,7 @@ import {
   SERVER_TYPE_ERROR,
   TYPE_UNHANDLED_ERROR,
   TYPE_UNHANDLED_REJECTION
-} from '../../shared/constants';
+} from '../../constants';
 
 export type SupportedErrorEvent = {
   id: number;
